+++ src/main/php/Param.php
@@ -50,20 +50,44 @@ class Param
      * returns name of param
      *
      * @return  string
+     * @since   3.0.0
+     */
+    public function name()
+    {
+        return $this->name;
+    }
+
+    /**
+     * returns name of param
+     *
+     * @return  string
+     * @deprecated since 3.0.0, use name() instead, will be removed with 4.0.0
      */
     public function getName()
     {
-        return $this->name;
+        return $this->name();
     }
 
     /**
      * returns value of param
      *
      * @return  string
+     * @since   3.0.0
+     */
+    public function value()
+    {
+        return $this->value;
+    }
+
+    /**
+     * returns value of param
+     *
+     * @return  string
+     * @deprecated since 3.0.0, use value() instead, will be removed with 4.0.0
      */
     public function getValue()
     {
-        return $this->value;
+        return $this->value();
     }
 
     /**
@@ -137,9 +161,21 @@ class Param
      * returns list of error ids
      *
      * @return  ParamError[]
+     * @since   3.0.0
+     */
+    public function errors()
+    {
+        return $this->errors;
+    }
+
+    /**
+     * returns list of error ids
+     *
+     * @return  ParamError[]
+     * @deprecated since 3.0.0, use errors() instead, will be removed with 4.0.0
      */
     public function getErrors()
     {
-        return $this->errors;
+        return $this->errors();
     }
 }