+++ src/main/php/filter/MailFilter.php
@@ -32,7 +32,7 @@ class MailFilter implements Filter
             return null;
         }
 
-        $value = $param->getValue();
+        $value = $param->value();
         $mailValidator = new MailValidator();
         if (!$mailValidator->validate($value)) {
             $param->addError($this->detectErrorId($value));