+++ src/main/php/filter/MonthFilter.php
@@ -42,7 +42,7 @@ class MonthFilter implements Filter
         }
 
         try {
-            return Month::fromString($param->getValue());
+            return Month::fromString($param->value());
         } catch (IllegalArgumentException $iae) {
             $param->addError('MONTH_INVALID');
         }