+++ src/test/php/ParamTest.php
@@ -21,7 +21,7 @@ class ParamTest extends \PHPUnit_Framework_TestCase
     public function returnsGivenName()
     {
         $param = new Param('foo', 'bar');
-        $this->assertEquals('foo', $param->getName());
+        $this->assertEquals('foo', $param->name());
     }
 
     /**
@@ -30,7 +30,7 @@ class ParamTest extends \PHPUnit_Framework_TestCase
     public function returnsGivenValue()
     {
         $param = new Param('foo', 'bar');
-        $this->assertEquals('bar', $param->getValue());
+        $this->assertEquals('bar', $param->value());
     }
 
     /**
@@ -84,7 +84,7 @@ class ParamTest extends \PHPUnit_Framework_TestCase
     public function hasEmptyErrorListByDefault()
     {
         $param = new Param('foo', 'bar');
-        $this->assertEquals([], $param->getErrors());
+        $this->assertEquals([], $param->errors());
     }
 
     /**
@@ -104,7 +104,7 @@ class ParamTest extends \PHPUnit_Framework_TestCase
     {
         $param = new Param('foo', 'bar');
         $error = $param->addError('SOME_ERROR', ['some' => 'detail']);
-        $this->assertEquals(['SOME_ERROR' => $error], $param->getErrors());
+        $this->assertEquals(['SOME_ERROR' => $error], $param->errors());
     }
 
     /**
@@ -128,7 +128,7 @@ class ParamTest extends \PHPUnit_Framework_TestCase
     {
         $param = new Param('foo', 'bar');
         $error = $param->addError(new ParamError('SOME_ERROR', ['some' => 'detail']));
-        $this->assertEquals(['SOME_ERROR' => $error], $param->getErrors());
+        $this->assertEquals(['SOME_ERROR' => $error], $param->errors());
     }
 
     /**